AUT-3749: Add `ReauthFailed` metric
Metric raised when the reauth journey has failed. This could be for many reasons, which is captured in a new dimension `FailureReason`.

These are raised when the audit event for failure is raised too. They will allow us to chart this metric in dashboards and establish a pattern of use that will could spot issues.