ATO-980: Moves parsing start request up
We now need pass the claim 'authenticated` from the start request
body into the other functions which will need us to parse this
request earlier on in the handler than previously. Therefore this
has now been moved further up and returns earlier if the request is
invalid.

This then means that we are no longer swallowing the potential JSON
exception when we "return generateApiGatewayProxyResponse(200 ..."
so we now need to handle this explicitly. I have adopted throwing
a runtime exception here as we do in similar cases throughout the
codebase