ATO-980: Update start service to use Authenticated from body
This updates the start service to use the authenticated claim from the
request body instead of using the shared session to determine this.
This removes the dependcy on the shared session store for this value.

There is some additional session logic in the validate session method which
we may want to revisit in future work