ATO-1049: Adjust logging when handling AIS exception
When AIS went down we triggered a P1 pager duty alarm (production-P1-ipv-handback-alarm) because we were logging at error level when handling AIS exceptions.

After this change we log at warn level if we are failing open, and error level if we are aborting on error. This will prevent sending the alalarm to pager duty when it is not P1.