BAU: Add a task to search all projects for a dependency.
This supports triage of vulnerability reports as it can be used in conjunction with dependencyInsight to determine what is causing the dependency to be included if it is transitive.  See the Security Vulnerabilites Runbook (https://govukverify.atlassian.net/wiki/spaces/LO/pages/4164649233/BAU+Daytime+Support+Hygiene+and+Optimisation+Rota#SEC-1%3A-Dependabot-Vulnerability-Alerts)

Also added a task to all projects that outputs all the resolved dependencies for all projects.