ATO-1088: Set Authenticated regardless of journey type
Currently we set the authenticated flag in the session here only
in the case of an auth only journey. This is a bit unusual as it
means that if on an identity journey we have to complete it for
our session to count as authenticated. We can fix this by setting
this value in the session before we redirect to IPV, therefore
setting in in all cases in which auth hand back a user