BAU: use different value in audit event
This pair had a key 'rpPairwiseId' and value userInfo.getClaim(‘rp_client_id’). rp_client_id is a claim added in AuthorisationHandler, but nowhere else. ClientId is already associated with the submitAuditEvent, and the internalPairwiseId is also already on the event as the subject of the user. So I’m assuming this was trying to add the rpPairwiseId, as rp_pairwise_id is a claim on the userInfo response from authentication.