ATO-980: Parse Start request earlier
We now need to extract the field `authenticated` from the start
request earlier so we need to move the parsing up earlier in the
handler. This will allow us to pass the field to the relevant
functions.

This also means we're no longer silently swallowing an error when
calling `generateApiGatewayProxyResponse(200 ...` meaning we now
have to explicitly handle the JSON exception from this callsite.
Instead now we catch this error and return a 400 with an appropriate
error messasge