BAU: Adds auth attempt extension to start handler test
The start handler access the Authentication attempts table via the
AuthenticationAttemptsService, however the test do not set up this
resource in local stack, leading to ResourceNotFound Exceptions when
attempting to run the test suite individually.

To fix this I have just added the extension for this table to the
test. This means the test suite can be ran by itself