ATO-980: Refactor logic to use authenticated claim
We can now refactor our logic to create a global flag which uses
the authenticated claim and the user profile to determine whether
a user counts as authenticated.

This requires some changes to the start service, such as splitting
off some of the logic from the verify session method into two small
methods. This tries to clarify the behaviour a bit whilst ensuring
we don't reintroduce the bug[1] this change was introduced to fix

[1]- https://github.com/govuk-one-login/authentication-api/commit/3a3752321071e6398b6ea7031bb58486d66b5016