ATO-1088: Removes setting authenticated in IPV callback and authcode
This removes the method saveSession from the auth code generation service,
which was respoonsible for updating a non doc-app session to be authenticated
and account state to exsisting. This is redundant as we now set both of these values[1][2]
in authentication callback handler once a user is handed back from authentication.

We can remove the method completley in IPV callback handler as isDocAppJourney
is always false. In authCode handler,we can replae this with calls to update the
session in the case of a doc app journey, as for a non doc-app journey we no longer
need to update the values that were previously being updated.

[1] - Setting isNewAccount based on Auth's userInfo response https://github.com/govuk-one-login/authentication-api/pull/5330/commits/d1a0d2e6d379775dbff84c7f2c14f4f1efbe5a9d
[2] - See commit 8e9316dbefe495e0d64d1c48c30492388c55ffa0