ATO-1104: Temporarily disable TTL configuration
To be able to modify the `attribute_name` field we firstly need
to disable[1] the TTL configurtion first to then be able to enable
the renamed attribute, which we will set to the exisiting attribute
`ttl`.

!NOTE-  Once this is merged, we must wait at least one hour before
merging the subsequent PR, which will re-enable the TTL config
with a new attribute. This is due to the underlying way dynamoDB
works[2]

[1] - https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-dynamodb-table-timetolivespecification.html
[2] - https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_UpdateTimeToLive.html#API_UpdateTimeToLive_RequestSyntax:~:text=It%20can%20take%20up%20to%20one%20hour%20for%20the%20change%20to%20fully%20process.%20Any%20additional%20UpdateTimeToLive%20calls%20for%20the%20same%20table%20during%20this%20one%20hour%20duration%20result%20in%20a%20ValidationException.