ATO-1104: Rename TTL attribute and turn back on
We have temporarily disabled the TTL configuration[1] on the table,
now we can re-enable this with the renamed attribute to reflect the
attribute we have been storing for the previous session entries. This
should have no affect on exisisting session and should eventually[2]
expire the built up sessions which will have already expired.

[1] - https://github.com/govuk-one-login/authentication-api/pull/5572
[2] -https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/TTL.html#:~:text=Items%20with%20valid%2C%20expired%20TTL%20attributes%20may%20be%20deleted%20by%20the%20system%20at%20any%20time%2C%20typically%20within%20a%20few%20days%20of%20their%20expiration