ATO-1118: add phone number to auth UserInfo response for all identity journeys
Orch needs access to a user's phone number only within identity journeys, so following the pattern of email and local_account_id, we add phone number as a claim for all identity journeys. This also adds a test to for low level trust journeys, so we can see that identity journeys add the specified set of claims.