ATO-975: Add current credential strength into the auth session
Taking the current credential strength from start passed from the frontend
and adding it into the auth session so it can be taken later on to be passed
on in the userinfo. The current credential field might not exist from the frontend
and therefore could be null. This is what we expect and we want to add null into the
auth session as it used later on in orch.