BAU: Move creation of api response out of service and into handler
The core responsibilities of a handler is to delegate to other services to deal with a request, and return a response. It's a bit easier to read if we can see what response we return in the handler, rather than putting this in a service.

This involved removing a test that was no longer relevant to the service (since as we're not constructing a json response here anymore, we're not throwing the relevant error here) and also removing an assertion for empty headers, since this was based on creating a mocked api response before.