BAU: Move emission of cloudwatch metric from service to handler
This helps the service method have more of a single responsibility and for it to be clearer in the handler what is going on when we handle a request.

This involved adding the cloudwatch metrics service to the handler constructor and out of the service constructor and adding a new test assertion.