BAU: update expired token test
ID token claims sets can no longer be created as expired from instantiation. Because of this, the shouldNotFailSignatureValidationIfIdTokenHasExpired
test failed as it was trying to create an ID as already expired. I've deleted this test because it is difficult to test without using the sleep method,
and the test is testing the functionality of the packages we rely on (nimbus jose and oidc sdk v2) rather than anything to do with our implementation.