AUT-3845: Enable snapstart for AIS stub
The AIS stub is causing acceptance tests to fail when run in parallel. This happened because the stub lambda cold started another instance during the run, meaning the response took longer than the client configured timeout..

Here we remove the provisioned concurrency for the stub lambda, replacing it with snapstart. We also switch the architecture for this lambda to arm64 for the snapstart support.

Co-authored-by: Ethan Mills <ethan.mills@digital.cabinet-office.gov.uk>