AUT-3872: Use the endpoint-lambda module
Previously, we had both `endpoint-lambda` and `endpoint-module`, both
of which had all the lambda-creation logic in them. As `endpoint-lambda`
was built from `endpoint-module`, we can now use it in place of the
duplicated logic within `endpoint-module`.

This should have been done a while ago. It is being done now, so that
the wait logic is only defined in one place.