BAU: Do not run excessive tests
We only need to run tests if they will actually test something that has
been changed.

If a test is not marked as a required status, we skip it via gha's
`paths` feature.

If a test is marked as required, we detect if any java files have been
changed via the new `call_get_changed_files.yml` workflow. If no java
files have been changed, we skip all the actual test steps. This is
necessary, because we *must* report a passed status from all required
checks: if we don't run the test, no status is reported, so merging is
blocked forever.