BAU: Update provider locks
Additionally, add a mechanism to keep the lockfiles up to date,
both via pre-commit (if someone updates versions.tf and forgets
to re-lock) and via a new script.

We've chosen to use `tfupdate` over the `terraform lock` command,
as we have so many lockfiles to keep updated, and it ends up as
a super grubby loop. `tfupdate` does the version resolution once
and takes significantly less time to process all our modules.