ATO-1147: use new IAM role with orch session access
The new role has already been deployed, and so we can safely switch over the old role of the lambda to the new role. We only use this role if the lambda needs access, which is when orch is not stubbed by auth. Don't turn on in prod yet - we will check if there is any downtime in integration just in case (we aren't expecting any, but we will check for defensiveness)