ATO-1243: Add a retry for getting Auth sessions from DynamoDB. Retry uses strongly consistent read which "returns a response with the most up-to-date data, reflecting the updates from all prior write operations that were successful" [1].
This change was introduced as a result of an incident which took place [2] whereby in a small percentage of cases, sessions were not being retrieved from DynamoDB shortly after they were written. By using strongly consistent read we can guarantee that the session will be retrieved if it's in the DB.

Note that strongly consistent read is twice as expensive as eventually consistent read (the default) which is why it's only used on the retry.
[1] https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/HowItWorks.ReadConsistency.html
[2] https://gds.slack.com/archives/C02HCLREVV5/p1732264067902419