ATO-1104: Comment out TTL config
This is needed due to a bug in the terraform provider[1]. This bug
causes any updates to the shared module to fail as terraform detects
a non-exisitent diff and tried to disable an already disabled ttl
configuration, resulting in a validation error. Merging this PR will
result in a no-op deployment of the shared module and allow any
subsquent deployments to continue.

[1]- https://github.com/hashicorp/terraform-provider-aws/issues/10304