ATO-1117: add decoder and encoder to SaltHelper
Also replace the private method in UserInfoService with this helper.

At the moment, we get salt from eg the UserProfile table, which it isn't encoded in. However, it is encoded in the authUserInfo response. Therefore, when we get it from there, we will need to decode it. Hence, also define a decoder.

Also adds a test that encoding and decoding generates the same byte array.