ATO-1252: Stop using OrchSessionService addOrUpdateSessionId in auth journey
This method is causing problems because it does too much. It is better to give more control to the handler so we can first build the Orch session we want, and then update it in the DB. This prevents multiple calls to DynamoDB when updating the Orch session fields (ie. we can remove the ifPresentOrElse(update, add) in the diff).

It would be nice to remove the addOrUpdateSessionId method entirely, but it is still used in the doc app journey for now.