ATO-1252: Create OrchSessionItem copy constructor
In AuthorisationHandler, we need to update the Orch session using an existing Orch session object. Without the copy constructor, we would have to mutate the existing Orch session object in order to create the new one. While this can work in this case, it is not good practice as it can introduce hard-to-find bugs. Using the copy constructor ensures we leave the original object unmodified.