BAU: fix policy creation errors when orch stubbed
var.orch_environment and var.orch_session_table_encryption_key_arn are
only used when var.is_orch_stubbed is false.

Previously, if these were left empty (as they should be when unused),
they both would be set to an empty string. This was okay for
var.orch_environment, but not for the encryption key arn: policy
creation failed because the arn was invalid.

Now, we only create the relevant `aws_iam_policy_document` data sources
and `aws_iam_policy` resources when var.is_orch_stubbed is false.

These policies are required by the
`ipv-processing-identity-role-with-orch-session-access` role, so we
now only attach them when they've been created, via a ternary.