BAU: New role module to reduce policy attachment
Previously, we created many policies, and then attached them all to the
individual roles.

This meant that we often bumped up against the 20 policy limit for a
role. It also took a long time to run the terraform apply, as it had to
check MANY policies for changes.

This change introduces a replacement for the lambda-role module, which
takes in a map of policy documents lists and combines them into a
reduced number of policies specific to the role.

eg.

```hcl
module "lambda-role-policy-reduction" {
  source = "./modules/lambda-role-policy-reduction"

  role_name = "my-role"
  environment = "dev"
  default_tags = var.default_tags
  vpc_arn = var.vpc_arn

  policy_documents_to_attach = {
    "dynamodb" = [
      data.aws_iam_policy_document.dynamodb_write_policy.json,
      data.aws_iam_policy_document.dynamodb_read_policy.json,
    ],
    "s3" = [
      data.aws_iam_policy_document.s3_list_policy.json,
      data.aws_iam_policy_document.s3_read_policy.json,
      data.aws_iam_policy_document.s3_write_policy.json,
    ],
    "sqs" = [
      data.aws_iam_policy_document.sqs_policy.json,
    ],
    "redis" = [
      data.aws_iam_policy_document.redis_read_policy.json,
      data.aws_iam_policy_document.redis_write_policy.json,
    ],
  }
}
```
This would create and attach 5 policies:
1. base: Contains the logging, networking and xray policies that are
   common to all roles.
2. dynamodb: Contains the dynamodb policies
3. s3: Contains the s3 policy
4. sqs: Contains the sqs policy
5. redis: Contains the redis policy

previously, this would have created and attached 11 policies to the role
ie. 1 more than the limit.