ATO-1117: add method to convert from ByteBuffer to byte array
There is enough steps that its worth having this helper. It is also very important to rewind immediately, as if we do not rewind, this will cause an side effect at the next read that the ByteBuffer position is at the end of the buffer, which we don't want as a side effect of this read.