AUT-3929: Warn when metrics would fail validation
The version bump of `aws-embedded-metrics` from v2 to v4 [1] introduces validation of metrics that thow errors on failure.

This commit brings in the same validation methods implemented by AWS [2] but instead of throwing an error it logs a warning with stacktrace.

The intention is to let this run in environments for a time period and assess any warnings logged before continuing with the version bump. The data gathered will help us make the nessessary modifications across orch and auth.

[1]: https://github.com/govuk-one-login/authentication-api/pull/4993
[2]: https://github.com/awslabs/aws-embedded-metrics-java/pull/119/files#diff-95ba60790f295fcb1d765c68bd8876c4722e13ef3fd5a34c6965b5fe1b393786R1