AUT-3902: Add performance tuning params for ticf cri lambda
These are deliberately looking up any overrides for the account
interventions lambda. For now, we want to scale these two lambdas up in
parallel. This will get readjusted when we start fully using ticf in
production, but for now we want to ensure that any scaling applies to
both lambdas.

Co-authored-by: Tom Whitwell <tom.whitwell@digital.cabinet-office.gov.uk>