BAU: Make P1 alarm deploy only when required
Previously, when a P1 alarm was not needed, this would deploy on top of
module.account_interventions.aws_cloudwatch_metric_alarm.lambda_error_cloudwatch_alarm
so the configuration of the alarm would be modified each time. Now, we
only deploy the P1 alarm when required, and it will always be created
with a specific name.