BAU: Remove need for replace statements in modules
This replace function is, apparently, non-deterministic. This causes
some attributes to be marked as 'known after apply' when they shouldn't.

This change adds a new variable, `endpoint_name_sanitized`, which is
required if `endpoint_name` contains a period. This variable is then
used in place of `endpoint_name` in the resources that were using the
`replace` function.