ATO-1099: Adds max age handling
Adds handling of sessions where the user has been on a max
age journey. If it is the same user (ie same internal subject
identifier) we copy their client sessions over. Otherwise
we don't. In future we will send back channel logouts
and emit an audit event in the negative case.

There is some slight weirdness here in the fact that we:

a) Ignore cases where we can't retrieve the previous session/
shared session. This is purposeful - We can't do anything
without the previous sessions so we just continue. The user
has already re-authenticated so there should be no interruption
to their journey

b) Nulling out of previous session id - again deliberate.
We don't want to keep trying to re-attach client sessions/or
emitting audit events when a user returns after a max age journey,
so this 'switches' this behaviour off after they have gone through
a max age journey. This value will only be re-set if they are prompted
to go through a new max age journey