BAU: Run SonarCloud analysis on any change
Previously, only changes to Java files would trigger a check. This is fine for most things.

In one PR though, it had java files in that were flagged as duplicates. The PR contained terraform changes too.

The java changes were removed and force pushed, but SonarCloud was not re-run (as it contained no Java file changes).

SonarCloud then had an outdated view of the PR and still reported duplication issues.