ATO-1245: Refactor session setting in start handler
The current implementation of setting up the session in
the start handler uses a method which we would like to
remove. It does a lot of logic on the auth session object
and is a bit gnarly. This logic doesn't need to be done
anywhere else outside of the start handler.

This refactor will also enable us to have an auth Session
object to pass to BuildUserContext, which will allow us
to add the AuthSession to the user context without it being
optional