BAU: Make the sendTokenRequest service method private as it isn't called directly by any production code.
This method was made public to allow easier testing but it doesn't need to be public as no production code calls the ReverificationResultService this way.  The service method that is called is getToken and is part of the services public API.

Use wiremock to allow the getToken method to be tested.