ATO-1099: Refactors session setting and handle max age
Bit of a large commit but refactors session setup in the
authorisation handler to make it a bit more logical and add the
max age session handling.

The main net difference here in existing behaviour is the fact
that now if there is not a shared session or not an orch session,
we generate new sessions for both. This seems logical considering
there shouldn't ever be a case in which we have a one but not the other.

This allows the code to be a lot simpler and we can handle our max
age expiry in a clearer, more readable way.

Co-authored-by: Bill Dunn <bill.dunn@digital.cabinet-office.gov.uk>