AUT-3691: Add test to check the JWKS endpoint serves the public key used to sign requests sent to IPV.
Use the Junit 5 System Stubs extension to change the environment for the lambda.  This allows the id of a key created in localstack to be used when testing the lambda and the JwksService.

Added a log4j2-test.xml config to allow easier reading of the logs when running tests.