AUT-3691: Add integration test for MFAResetAuthorizeHandler.
The integration test checks how the lambda integrates with KMS, redis, SQS and CloudWatch.

+ Integration with KMS is checked to confirm the expected keys are being utilised.

To facilitate this testing the KmsKeyExtension was amended to record the keyId of the created key. Because localstack KMS doesn't allow key usage to be checked an interceptor was added to the KmsClient created in the KmsConnectionService for ue with localstack.  This interceptor simply records the latest key use for a signing operation.  This can be extended when checks are required for multiple keys and/or operations.

+ Integration with redis is checked to confirm the expected state is stored.

The existing RedisExtension is used.

+ Integration with SQS is checked to confirm events are published for TxMa.

The SQSQueueExtension is used.  Localstack is very slow at publishing events so the check for the TxMA event is retried, using awaitility, until the event can be read.  This is not a scalable approach and its value will be evaluated on overall test suite run times.  An alternative approach could be to add an interceptor to the SQS client.

+ Integration with CloudWatch is checked to confirm the expected metrics will be created.

Localstack does not automatically created CloudWatch metrics from CloudWatch logs so the logs are checked instead of looking up the metric.  This is acceptable as we do not need to test that AWS correctly processed log messages we just need to check that we have started the process correctly.

The MetricsLogger class is very difficult to configure such that a lambda class being run outside of the localstack container will log to localstack.  To work round this issue a test double of the MetricsLogger is used.  The test double uses the CloudWatch Logs API rather than the MetricsLogger but exposes the same interface as the MetricsLogger.  We can then check the required logs are produced.  The test double is very simple and can be extended as we find more complex test scenarios.

The logs service is enabled in the docker compose used locally and in the pre-merge-checks github workflow.