AUT-3814: Configure AWS credentials
For each supported environment, get and export AWS credentials using the existing script. The conditionals may seen to detailed for now, but these will be populated later with env specific variables (like clientIds and KMS key ARNs).

We need an AWS connection to get the data needed to calculate the internal common subject identifier and to get the keys needed to sign the auth token.