ATO-1117: add temporary logging to verify values
In a previous PR, I had assumed orchSession.getInternalCommonSubjectId() would be a defined string, but it seems this is not the case. When I tried to use it to get from the dynamodb table, it threw the error Supplied AttributeValue is empty a few times in 4 hours. Here, I add a small subset of the logs I added there to see what is going on.