AUT-3894: Clarify names to show auth ownership.
Changed the name of the IPV public key used by Auth (frontend-api) to make it clear that this is specifically for the auth components and that orchestration components use a different key called ipv_auth_public_encryption_key.

Cleaned up the dev and dev-overrides file so that the public IPV keys are provided in the overrides file to align with how it is done in the other environments.

Changed the name of the environment variable that provides the MFAResetAuthorization handler with the IPV public encrypting key to reflect the key name.  This environment variable is specific to the lambda and so does not need the `auth_frontend_api_to_` prefix used in terraform to differentiate it from the equivalent orchestration variable.