ATO-1117: check if rpPairwiseId is not blank
We have discovered that in our new structure, it is possible for a user to have an undefined internalCommonSubjectId here. In this cases, it was because the user had skipped authenticationCallbackHandler. However, there were 2 cases where internalCommonSubjectId was defined and matched the calculated value, but rpPairwiseId didn't. With this log, I want to check if the values actually don't match, or if the value is just undefined.