ATO-1117: wrap getting authUserInfo in a blank check for ICSID
This PR was previously merged without this commit, and it errored in a small percentage of cases. This was when users would try to log in (thus creating a session), but skip AuthenticationCallbackHandler (most likely by clicking in their history) and going straight to IPV. This meant their session had no sub data, including ICSID, and therefore also no authUserInfo data. This meant this code had an uncaught internal server error. This commit fixes this issue.