BAU: Parse max_age if integer in request object
Our previous code would assume that the max_age would be included
as a string claim inside the request object and otherwise
throw a ParseException if not a string. This seems odd as it
is plausible RPs will include max_age as an integer claim inside the
request object, and we should support both.

This is not a problem for the queryParams as this is passed as a
string to the handler.