BAU: small refactor of handler method
Here we can now refactor the getMaxAge method as we
have persisted the value to the top level
authRequest object and therefore we no longer need
to pull the value out of the request object every time.

This also includes some test changes to add invalid_request
tests for negative max_age in the integration tests and remove
them from the unit tests. This was because we're mocking the
validators in the unit tests so made it hard to test these cases
properly