BAU: create transitional modules
Introduce two modules that shadow the existing ones, but which make
live migration from PC to Snap Start much easier.

Rather than using the lambda alias as the qualifier for provisioned
concurrency, instead use the new version of the function.

This fixes an annoying ordering issue when switching from PC to
Snap Start: terraform tries to update the alias before the
provisioned concurrency configuration is removed, triggering a
ResourceConflictException, as it was attempting to create a state where
both PC and Snap Start were enabled at the same time.

Now, we create (or destroy) the PC configuration, then reassign the
alias.

THESE MODULES SHOULD NOT BE USED TO DEPLOY PROVISIONED CONCURRENCY.