AUT-3894: Corrections to storage token claims.
Testing in Staging showed that we were not passing the correct audience
in the storage token JWT to allow EVCS to process it.  The audience is
an array containing IPV and EVCS audience uris now.

The issuer claim for the storage token was also missing from the
embedded JWTs claim set.  This has been added and set to the correct
issuer id which for auth is currently the url of the auth frontend.