AUT-3895: Tidying up environment files.
We have two terraform files that provide environment specific values.
In addition, some of the environment variables are shared with
orchestration and some are just for auth.  This commit does a minor
re-organisation of the variables so they are in a section where their
use is clearer and grouped together by feature.

Introduced a tfvar for the issuer claim used provided to IPV.  This is
highlighted as non-standard by having the _IPV suffix.  IPV currently
expect a string rather than the standard URI.  EVCS expect a URI so
the name of the variable we use for this can be generic for auth as we
would use the URI as our issuer claim in all request claim sets.