AUT-3895: Align dev config layout to staging.
To make it easier to compare the environment files the dev files have
been re-organised to follow the same ordering as staging.  Some vars
have been moved between files for dev and any vars that are defined in
staging but not in dev have been added as place holder comments so that
the files align when viewed next to each other.

Sections have been introduced to show which vars are common and used by
both auth and orchestration; which vars may no longer be needed; and
which are specific to an environment and not seen elsewhere.