AUT-3916: Clarification to key id provided in JWKS
AWS calls the id of a key a keyId but in the JWKS the key id is
abbreviated to kid.  This is confusing so the name of the expected key
id in the test was changed to make it clear that its not the AWS keyId
and what is in the kid.