AUT-3916: Replace localstack interceptor.
After review we decided to not pursue the approach of adding a
localstack interceptor to the KMS connection as this is only a
requirement for testing.  Although we have production code that checks
if localstack is being used instead of AWS and specifically configures
the KMS connection for localstack we don't like having this code in
our code base and so will not build on it.

The alternative approach is to use Wiremock to proxy calls to localstack
and then query the requests proxied to check the correct information, in
this case key id, is specified in the requests to KMS.

This commit cleans up the ConfigurationService and KmsConnectionService
by removing reference to the KMS localstack interceptor.