AUT-3916: Remove tests for cloudwatch metrics.
After review we decided to not pursue the approach of testing that
metrics are successfully published due to the MetricsLogger being very
difficult to configure to work with localstack.  This precluded us
from using the wiremock interceptor approach.

We also found that the MetricsLogger fails silently which is something
we will investigate separately.

Fixed issue in the MfaResetAuthorizeHandlerIntegrationTest where it was
using a static method on the WireMock class to verify a result rather
than the instance created in the test.