BAU: terraform validate before dev deploy
If terraform validate fails, the deploy in CI will fail. therefore it
makes sense to validate it here. As the validation is done at the same
time as the java build, it will not slow down this workflow.

The original terraform validate step in `pre-merge-checks-terraform.yml`
has been moved to a `workflow_call` workflow - allowing the same logic
to be run both in pre-merge-checks, and in the dev deploy workflow.