AUT-4030: when deploying dev env build into tf dir
Previously, the `buildZip` tasks would build the zip into gradle's
standard `build` directory. This is not the same location they are
located when running in CI.

This change updates the `buildZip` tasks to build the zip into the
`ci/terraform/<service>/artifacts` directory when the `devDeployBuild`
property is set. This property is set in the `deploy-dev.sh` script, and
should never need to be set manually.