BAU: reuse terraform data dir between branch runs
It's annoying to have to keep re-init'ing all the terraform modules
every time terraform is run. This change makes it so that the terraform
data dir is retained between runs of the same branch, so that the
`init` step takes less time on subsequent runs.

The data dirs can be removed by passing `-c` or `--clean` to the script.

These data dirs are stored in `${TMPDIR}/authentication-api-tf/${b64_branch_name}`,
where `${b64_branch_name}` is the base64 encoded branch name.

They will be removed on every reboot, so there's no need to worry about
regular cleanup.