BAU: Make the PR creation experience more pleasant
PR #5692 modified the template to add an orchestration-specific section
of the template. It's quite awkward to have to delete a huge chunk of
the file each time you want to create a PR.

We now have two 'proper' templates, one for auth, one for orch. The
'main' template has been updated to just be a picker, to redirect users
on to the correct template. You do have to click through to the Preview
tab to actually use the links, but it's less clunky than deleting
multiple sections of the file.

`gh` users can use the `--template` flag to pre-select the correct template
when creating a PR (or choose it when prompted in the wizard).